+++ eflo-controller-20221215/src/Models/CreateClusterRequest/nodeGroups/hyperNodes/dataDisk.php
@@ -1,0 +1,90 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Eflocontroller\V20221215\Models\CreateClusterRequest\nodeGroups\hyperNodes;
+
+use AlibabaCloud\Dara\Model;
+
+class dataDisk extends Model
+{
+    /**
+     * @var string
+     */
+    public $category;
+
+    /**
+     * @var bool
+     */
+    public $deleteWithNode;
+
+    /**
+     * @var string
+     */
+    public $performanceLevel;
+
+    /**
+     * @var int
+     */
+    public $size;
+    protected $_name = [
+        'category' => 'Category',
+        'deleteWithNode' => 'DeleteWithNode',
+        'performanceLevel' => 'PerformanceLevel',
+        'size' => 'Size',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->category) {
+            $res['Category'] = $this->category;
+        }
+
+        if (null !== $this->deleteWithNode) {
+            $res['DeleteWithNode'] = $this->deleteWithNode;
+        }
+
+        if (null !== $this->performanceLevel) {
+            $res['PerformanceLevel'] = $this->performanceLevel;
+        }
+
+        if (null !== $this->size) {
+            $res['Size'] = $this->size;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Category'])) {
+            $model->category = $map['Category'];
+        }
+
+        if (isset($map['DeleteWithNode'])) {
+            $model->deleteWithNode = $map['DeleteWithNode'];
+        }
+
+        if (isset($map['PerformanceLevel'])) {
+            $model->performanceLevel = $map['PerformanceLevel'];
+        }
+
+        if (isset($map['Size'])) {
+            $model->size = $map['Size'];
+        }
+
+        return $model;
+    }
+}